refactor : chat-content에서 미흡했던 에러핸들림을 추가함
why : 사용자 혹은 개발자가 에러에 잘 대처할 수 있도록 핸들링함
