+++ client/src/presentation/components/temp/chat-content.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
 import { MessageCard } from "./message-card";
-import { useMessages } from "../../contexts/messages-context";
-
+import { useMessages } from "../../../contexts/messages-context";
+import { Message } from "../../../core/entity/message";
 const ChatContentWrapper = styled.section`
   height: 100%;
   width: 100%;
@@ -13,18 +13,27 @@ const ChatContentWrapper = styled.section`
 `;
 
 export const ChatContent: React.FC = () => {
-  const messages = useMessages();
-  return (
-    <ChatContentWrapper>
-      {messages.map(message => (
-        <MessageCard
-          key={message.id}
-          name={message.name}
-          imageSrc={message.imageSrc}
-          contents={message.contents}
-          timestamp={message.timestamp}
-        />
-      ))}
-    </ChatContentWrapper>
-  );
+  let messages: Message[] | null;
+
+  try {
+    messages = useMessages();
+  } catch (error) {
+    alert(error.message);
+    messages = null;
+  }
+
+  function messageList(): React.ReactNode {
+    if (!messages) return <></>;
+    return messages!.map(message => (
+      <MessageCard
+        key={message.id}
+        name={message.name}
+        imageSrc={message.imageSrc}
+        contents={message.contents}
+        timestamp={message.timestamp}
+      />
+    ));
+  }
+
+  return <ChatContentWrapper>{messageList()}</ChatContentWrapper>;
 };
