+++ client/src/presentation/components/temp/message-section-content.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { ChatContent } from "./chat-content";
 import { ChatInputBox } from "./chat-input-box";
-import { MessageContextProvider } from "../../contexts/messages-context";
+import { MessageContextProvider } from "../../../contexts/messages-context";
 
 const MessageSectionContentWrapper = styled.section`
   width: 100%;
